:sparkles: add Token

+++ contracts/Token.sol
@@ -4,7 +4,7 @@ pragma solidity ^0.8.6;
 import "@openzeppelin/contracts/token/ERC20/ERC20.sol";
 
 contract Token is ERC20 {
-    constructor(uint256 initialSupply) ERC20("Treeb", "TREEB") {
+    constructor(uint initialSupply) ERC20("Treeb", "TREEB") {
         _mint(msg.sender, initialSupply);
     }
 }
